+++ redux-starter/src/components/app.jsx
@@ -1,16 +1,20 @@
 import ReactDOM from 'react-dom';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import Home from './home';
+import Home from './homePage/home';
 import About from './about';
 import Contact from './contact';
+import Navbar from './navbar';
 
 const App = () => (
   <BrowserRouter>
-    <Switch>
-      <Route exact path="/" component={Home} />
-      <Route path="/about" component={About} />
-      <Route path="/contact" component={Contact} />
-    </Switch>
+    <div>
+      <Navbar />
+      <Switch>
+        <Route exact path="/" component={Home} />
+        <Route path="/about" component={About} />
+        <Route path="/contact" component={Contact} />
+      </Switch>
+    </div>
   </BrowserRouter>
 );
