+++ redux-starter/src/components/homePage/home.jsx
