+++ redux-starter/src/stylesheets/main.sass
@@ -1,6 +1,3 @@
-.d-flex
-  display: flex
-.justify-content-between
-  justify-content: space-between
-.align-items-center
-  align-items: center
+@import variables
+@import utilities
+@import page
