+++ redux-starter/src/stylesheets/utilities.sass
@@ -1,0 +1,54 @@
+// Layout Classes
+.d-block
+  display: block
+.d-flex
+  display: flex
+.justify-content-center
+  justify-content: center
+.justify-content-between
+  justify-content: space-between
+.justify-content-around
+  justify-content: space-around
+.justify-content-start
+  justify-content: flex-start
+.justify-content-end
+  justify-content: flex-end
+.align-items-center
+  align-items: center
+.align-items-start
+  align-items: flex-start
+.align-items-end
+  align-items: flex-end
+.flex-fill
+  flex: 1
+.flex-wrap
+  flex-wrap: wrap
+.flex-direction-column
+  flex-direction: column
+
+// Show/Hide Classes
+.hide
+  display: none
+.show
+  display: block
+
+// Text Classes
+.text-white
+  color: $white
+.text-black
+  color: $black
+
+// Cursor Classes
+.c-pointer
+  cursor: pointer
+
+// Font Weights Generator
+@for $i from 3 through $font-weights
+  .fw-#{$i*100}
+    font-weight: #{$i*100}
+
+// Font Sizes Generator
+@for $i from 10 through $font-sizes
+  .fs-#{$i}
+    font-size: #{$i}px
+    line-height: 1.5
