+++ redux-starter/src/stylesheets/variables.sass
@@ -1,0 +1,7 @@
+// Colors
+$white: #fff
+$black: #000
+
+// Generator Constants
+$font-weights: 9
+$font-sizes: 36
